added relevant changes for spring cloud

+++ src/main/java/com/cipherLab/user/Controller/UserController.java
@@ -3,7 +3,11 @@ package com.cipherLab.user.Controller;
 import com.cipherLab.user.Entity.UserEntity;
 import com.cipherLab.user.Exception.UserNotFoundException;
 import com.cipherLab.user.Repository.UserRepository;
+import com.cipherLab.user.constants.UserConstant;
+import com.cipherLab.user.dto.ResponseDto;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
@@ -16,17 +20,27 @@ public class UserController {
     private UserRepository userRepository;
 
     @GetMapping("/users")
-    public List<UserEntity> getAll() {
-        return userRepository.findAll();
+    public ResponseEntity<List<UserEntity>> getAll() {
+        var users = userRepository.findAll();
+        return ResponseEntity.status(HttpStatus.OK).body(users);
     }
 
     @GetMapping("/users/{id}")
-    public UserEntity getUserById(@PathVariable Long id) {
-        return userRepository.findById(id).orElseThrow(() -> new UserNotFoundException(id));
+    public ResponseEntity<Optional<UserEntity>> getUserById(@PathVariable Long id) {
+        var user = userRepository.findById(id);
+        if (user.isEmpty()) throw new UserNotFoundException(id);
+        else return ResponseEntity.status(HttpStatus.OK).body(user);
+    }
+
+    @GetMapping("/users/byCellNumber/{cellNumber}")
+    public ResponseEntity<List<UserEntity>> getBookByCellNumber(@PathVariable String cellNumber) {
+        var user = userRepository.findByCellNumber(cellNumber);
+        if (user.isEmpty()) throw new UserNotFoundException(cellNumber);
+        else return ResponseEntity.status(HttpStatus.OK).body(user);
     }
 
     @PutMapping("/users/{id}")
-    public UserEntity updateUserById(@RequestBody UserEntity newUserEntity, @PathVariable Long id) {
+    public ResponseEntity<ResponseDto> updateUserById(@RequestBody UserEntity newUserEntity, @PathVariable Long id) {
         Optional<UserEntity> userOptional = userRepository.findById(id);
         if (userOptional.isEmpty()) {
             throw new UserNotFoundException(id);
@@ -35,18 +49,25 @@ public class UserController {
             oldUser.setName(newUserEntity.getName() != null ? newUserEntity.getName() : oldUser.getName());
             oldUser.setAge(newUserEntity.getAge() != null ? newUserEntity.getAge() : oldUser.getAge());
             newUserEntity.setLastChangeTs(LocalDateTime.now());
-            return userRepository.save(oldUser);
+            userRepository.save(oldUser);
+            return ResponseEntity.status(HttpStatus.OK).body(new ResponseDto(HttpStatus.OK, UserConstant.MESSAGE_200));
         }
     }
 
     @PostMapping("/users")
-    public UserEntity createUser(@RequestBody UserEntity newUserEntity) {
+    public ResponseEntity<ResponseDto> createUser(@RequestBody UserEntity newUserEntity) {
         newUserEntity.setLastChangeTs(LocalDateTime.now());
-        return userRepository.save(newUserEntity);
+        userRepository.save(newUserEntity);
+        return ResponseEntity.status(HttpStatus.CREATED).body(new ResponseDto(HttpStatus.CREATED, UserConstant.MESSAGE_201));
     }
 
     @DeleteMapping("/users/{id}")
-    void deleteUser(@PathVariable Long id) {
-        userRepository.deleteById(id);
+    ResponseEntity<ResponseDto> deleteUser(@PathVariable Long id) {
+        var user = userRepository.findById(id);
+        if (user.isEmpty()) throw new UserNotFoundException(id);
+        else {
+            userRepository.deleteById(id);
+            return ResponseEntity.status(HttpStatus.OK).body(new ResponseDto(HttpStatus.OK, UserConstant.MESSAGE_200));
+        }
     }
 }
